Added helped to manage process execution
This commit adds a new file, CmdProcess.cs, which contains the implementation of a class that manages command line processes. The class has methods for creating, starting and killing processes as well as events for when the process starts and exits or when data is received from stdout or stderr.