Added method to update GUI components
This commit refactors the RunnerForm.cs file by adding a new method UpdateComponents() that updates the start and stop buttons, as well as the file to execute label. The CmdProcess class is also instantiated in the constructor of RunnerForm. These changes improve the user interface and make it more responsive.