Changed settings file location
This commit adds the ability to save program settings to a file. The `Program.cs` file was modified to use the `System.IO` namespace and create a new path for the settings file in the user's application data folder. The `Settings.cs` file was also modified to create the directory if it doesn't exist before creating or deleting the configuration file.