Added event listeners to update settings
This commit adds new methods to the SettingsForm class that allow for updating the GUI components and their width, as well as adding functionality to update settings when text is changed in the FileToExecuteTextBox or ArgumentTextBox. Additionally, a SaveButton_Click method was added to set settings as updated and close the form, while an OnClosed method was added to save or restore settings based on whether they were updated or not. Finally, event handlers were added for various controls such as FileToExecuteLabel.Click and SaveButton.Click.