Added error button to show/hide errors
This commit adds a new button that allows the user to toggle between showing and hiding error messages. The button is enabled only when there are errors to display. Additionally, the code has been refactored to include a new method `ShowOutputText` which updates the GUI by invoking the correct thread and bringing either the output or error text box to the front depending on which one was last shown. Finally, some minor changes have been made to improve readability and consistency in naming conventions.