Changed how to handle process output
This commit adds a new class called OutputReceivedEventArgs that inherits from EventArgs. It also adds a delegate called OutputReceivedEventHandler that takes an object sender and an instance of the newly created class as arguments. The ProcessHelper class now uses this delegate to handle process output instead of the DataReceivedEventHandler. Two new events are added, ErrorDataReceived and OutputDataReceived, both using the new delegate. Additionally, two threads are created for reading error and output data from the process respectively. Finally, in RunnerForm.cs, the Process_Output method is updated to use the new OutputReceivedEventArgs instead of DataReceivedEventArgs when handling process output.