Unified some class and method names
The commit refactors the code by renaming classes and methods to be more descriptive. The `OutputReceivedEvent` class is renamed to `RunnerEvent`, with its corresponding delegate being renamed as well. The `Process_Output` method in the `RunnerForm` class is also renamed to include "Runner" in its name, making it clearer what it does. Additionally, some unused files are removed from the project file.