Added helper for automatic updated
This commit adds two new classes, Updater and UpdaterEvent, which are used to update the program from GitHub. The Updater class contains methods for downloading data from GitHub, parsing it, and downloading updates if necessary. The UpdaterEvent class defines an event handler that is called when an update has been completed.