Improved output and error text clearing
Previously, the code cleared the output text box when clean was true. Now, it clears either the output or error text box depending on whether an error occurred. The UpdateOutputText method has been updated to take a new parameter that specifies whether to clear the error or output text box.