Moved and fixed update process workflow
This commit refactors the Updater.cs file to include functionality for backing up and replacing files. It sets a temporary path to write the new file, deletes any previous backups of the new file, backs up the current executable file, moves the existing new file to a backup location if it exists, moves the new file from tmp to its final location, and sets the path of downloaded files.