Added options to save settings with delay
This commit refactors the `Settings` class to allow for delayed saving of settings. A new private method `SaveInternal` was added to handle the actual saving, while a public method `Save` was modified to accept an optional boolean parameter `delay`. When set to true, this parameter starts a new thread that waits for 1 second before calling `SaveInternal`. If a previous save thread is still running, it is aborted before starting the new one.