Added save counter text in debug mode
This commit adds new features to the RunnerForm. In debug mode, it updates the counter text when the counter is increased and shows the current value of the counter. It also adds a CounterText label to display this information. The Settings class now has a Counter property that tracks how many times settings have been saved, and an event handler that invokes when this value is increased.